api/pyproject.toml
[tool.poetry]
name = "purplesector-fullstackcodingchallenge"
version = "0.1.0"
description = ""
authors = ["Alex Florea <alex.florea@ps.tech>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
daphne = "^4.1.2"
channels = "^4.1.0"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-filter = "^24.3"
psycopg = "^3.2.3"
whitenoise = "^6.8.2"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
coverage = "^7.6.4"
debugpy = "^1.8.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
